Move bookmark migration to feature domain module
